requirements.txt
pymongo>=4.0.0
python-dateutil>=2.8.0
tqdm>=4.65.0
neo4j>=5.20.0
python-dotenv>=1.0.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
rich>=13.7.0
scipy>=1.10.0
